Moved extractor-oriented code from Extractors to Extractor.
Warning: 8 ignores
Extractors: moved extractField, plural, attribute, optionalAttribute, optional, logger, translations, translateMemberName, extractChildren to Extractor.
