+++ src/main/scala/com/phasmidsoftware/xml/Extractor.scala
@@ -1,6 +1,12 @@
 package com.phasmidsoftware.xml
 
-import scala.util.{Failure, Try}
+import com.phasmidsoftware.core.Utilities.show
+import com.phasmidsoftware.core.XmlException
+import com.phasmidsoftware.xml.Extractors.{extractOptional, extractSingleton}
+import org.slf4j.{Logger, LoggerFactory}
+import scala.collection.mutable
+import scala.util.matching.Regex
+import scala.util.{Failure, Success, Try}
 import scala.xml.{Node, NodeSeq}
 
 /**
@@ -34,13 +40,57 @@ trait Extractor[T] {
      * @tparam P the type of the Extractor we wish to return.
      * @return an Extractor[P].
      */
-    def mapTo[P >: T]: Extractor[P] = (node: Node) => self.extract(node)
+    private def mapTo[P >: T]: Extractor[P] = (node: Node) => self.extract(node)
 }
 
 /**
  * Companion object to Extractor.
  */
 object Extractor {
+
+    /**
+     * Method to yield a Try[P] for a particular child or attribute of the given node.
+     *
+     * NOTE: Plural members should use extractChildren and not extractField.
+     *
+     * NOTE: ideally, this should be private but is used for testing and the private method tester is struggling.
+     *
+     * @param field the name of a member field:
+     *              if the (text) content of the node, then the field should be "$";
+     *              if a singleton child, then field is as is;
+     *              if an attribute, then field should begin with "_";
+     *              if an optional child, then field should begin with "maybe".
+     * @param node  a Node whence field is to be extracted.
+     * @tparam P the type to which Node should be converted [must be Extractor].
+     * @return a Try[P].
+     */
+    def extractField[P: Extractor](field: String)(node: Node): Try[P] = doExtractField[P](field, node) match {
+        case _ -> Success(p) => Success(p)
+        case m -> Failure(x) =>
+            x match {
+                case _: NoSuchFieldException => Success(None.asInstanceOf[P])
+                case _ =>
+                    val message = s"extractField ($m): field '$field' from node:\n    {${show(node)}}"
+                    logger.warn(s"$message caused by $x")
+                    Failure(XmlException(message, x))
+            }
+    }
+
+    /**
+     * Method to extract child elements from a node.
+     *
+     * @param member the name of the element(s) to extract, according to the construct function (typically, this means the name of the member in a case class).
+     * @param node   the node from which we want to extract.
+     * @tparam P the (MultiExtractor-able) underlying type of the result.
+     * @return a Try[P].
+     */
+    def extractChildren[P: MultiExtractor](member: String)(node: Node): Try[P] = {
+        val w = translateMemberName(member)
+        val nodeSeq = node \ w
+        if (nodeSeq.isEmpty) logger.info(s"extractChildren: no children found for child $w (for member $member) in ${show(node)}")
+        implicitly[MultiExtractor[P]].extract(nodeSeq)
+    }
+
     /**
      * Method to create an Extractor[T] which always fails.
      *
@@ -49,21 +99,62 @@ object Extractor {
      */
     def none[T]: Extractor[T] = (_: Node) => Failure(new NoSuchElementException)
 
+
+    val translations: mutable.HashMap[String, String] = new mutable.HashMap[String, String]()
+
+    private def translateMemberName(member: String): String =
+        translations.getOrElse(member,
+            member match {
+                case plural(x) => x
+                case _ => translations.getOrElse(member, member)
+            })
+
+    private def doExtractField[P: Extractor](field: String, node: Node) = field match {
+        // NOTE special name for the (text) content of a node.
+        case "$" => "$" -> extractText[P](node)
+        // NOTE attributes must match names where the case class member name starts with "_"
+        case attribute("xmlns") => "attribute xmlns" -> Failure(XmlException("it isn't documented by xmlns is a reserved attribute name"))
+        case optionalAttribute(x) => s"optional attribute: $x" -> extractAttribute[P](node, x, optional = true)
+        case attribute(x) => s"attribute: $x" -> extractAttribute[P](node, x)
+        // NOTE child nodes are extracted using extractChildren, not here.
+        case plural(x) => s"plural:" -> Failure(XmlException(s"extractField: incorrect usage for plural field: $x. Use extractChildren instead."))
+        // NOTE optional members such that the name begins with "maybe"
+        case optional(x) => s"optional: $x" -> extractOptional[P](node \ x)
+        // NOTE this is the default case which is used for a singleton entity (plural entities would be extracted using extractChildren).
+        case x => s"singleton: $x" -> extractSingleton[P](node \ x)
+    }
+
+    private def extractAttribute[P: Extractor](node: Node, x: String, optional: Boolean = false): Try[P] =
+        (for (ns <- node.attribute(x)) yield for (n <- ns) yield implicitly[Extractor[P]].extract(n)) match {
+            case Some(py :: Nil) => py
+            case _ if optional => Failure(new NoSuchFieldException)
+            case _ => Failure(XmlException(s"failure to retrieve unique attribute $x from node ${show(node)}"))
+        }
+
     /**
-     * Method (if needed) to uncurry a 6-level curried function.
-     * Not used currently.
-     *
-     * @param f the original function.
-     * @tparam T1 type of parameter 1.
-     * @tparam T2 type of parameter 2.
-     * @tparam T3 type of parameter 3.
-     * @tparam T4 type of parameter 4.
-     * @tparam T5 type of parameter 5.
-     * @tparam T6 type of parameter 6.
-     * @tparam R  the type of the result.
-     * @return a function of type (T1, T2, T3, T4, T5, T6) => R
+     * Regular expression to match a plural name, viz. .....s
      */
-    def uncurry6[T1, T2, T3, T4, T5, T6, R](f: T1 => T2 => T3 => T4 => T5 => T6 => R): (T1, T2, T3, T4, T5, T6) => R = (t1, t2, t3, t4, t5, t6) => f(t1)(t2)(t3)(t4)(t5)(t6)
+    val plural: Regex = """(\w+)s""".r
+
+    /**
+     * Regular expression to match an attribute name, viz. _.....
+     */
+    val attribute: Regex = """_(\w+)""".r
+
+    /**
+     * Regular expression to match an optional attribute name, viz. &#95;&#95;.....
+     * With an optional attribute, it will have a default value that does not need to be overridden.
+     */
+    val optionalAttribute: Regex = """__(\w+)""".r
+
+    /**
+     * Regular expression to match an optional name, viz. maybe....
+     */
+    val optional: Regex = """maybe(\w+)""".r
+
+    private def extractText[P: Extractor](node: Node): Try[P] = implicitly[Extractor[P]].extract(node)
+
+    val logger: Logger = LoggerFactory.getLogger(Extractor.getClass)
 }
 
 /**
