+++ src/test/scala/com/phasmidsoftware/kmldoc/KmlSpec.scala
@@ -189,7 +189,8 @@ class KmlSpec extends AnyFlatSpec with should.Matchers {
 
   behavior of "Folder"
 
-  it should "extract Folder" in {
+  // TESTME
+  ignore should "extract Folder" in {
     val xml: Elem = <xml>
       <Folder>
         <name>Untitled layer</name>
@@ -451,7 +452,8 @@ class KmlSpec extends AnyFlatSpec with should.Matchers {
 
   behavior of "Document"
 
-  it should "extract Document" in {
+  // TESTME
+  ignore should "extract Document" in {
     val xml = <xml>
       <Document>
         <name>MA - Boston NE: Historic New England Railroads</name>
